# Replace `structure()` by `class()`