--- a/NEWS.md
+++ b/NEWS.md
@@ -1,5 +1,7 @@
 # xml2 (development version)
 
+* Small speedup for `xml_find_all()` (@mgirlich, #393).
+
 # xml2 1.3.4
 
 * Fixes for R CMD check problems.